+++ src/main/kotlin/io/testaxis/intellijplugin/toolwindow/projectstatistics/ProjectStatisticsTab.kt
@@ -1,7 +1,0 @@
-package io.testaxis.intellijplugin.toolwindow.projectstatistics
-
-import com.intellij.ui.components.Label
-
-class ProjectStatisticsTab {
-    fun content() = Label("Coming soon")
-}